add host attribute to allow the provider to use multiple Corellium instances
To allow the provider to be used by multiple Corellium instances, it
should support receive the `host` property to reach the right one.
Before this change, we could only reach the hard-coded instance's
address defined in the code.

If no values is passed to `host` attribute, the `app.corellium.com`
is set as the default one, but it can be overridden by
`CORELLIUM_API_HOST` env variable. The env variable WILL NOT
override the `host` when a its values is not the default one.